requirements.txt
numpy==2.2.1

mysql-connector-python~=9.1.0